Downgrade i18n from 0.9.5 to 0.9.1
Pins i18n to < 0.9.3 as this version introduces a bug that prevents the
fallback (based on the locale parent e.g. "fr" for "fr-BE") translations
from being stored in memory if they are not explicitly listed in
available_locales. (As yet unfixed in later versions, aside from being
able to switch the feature off by disabling enforce_available_locales.)

This results in calls to `I18n.translate` or `I18n.localize` where
static translations - e.g. date and time information - either raising an
error or an error message instead of the expected array of translations.