Strip AdminUserController#index query param
Leading and trailing whitespace is included in the search query,
preventing a search matching records in some cases.

The situation where this was noticed was pasting an email address –
presumably from a mail client – which included a leading whitespace.

Implemented using `try` as the `query` param can often be `nil`, so
`try` prevents a `NoMethodError` when not supplying `query`.

Fixes https://github.com/mysociety/alaveteli/issues/4855