Move EIR only message to top of page, give precedence over foi_no tag
We only want to show one of these messages at a time as only
being able to request EIR information implies not being
bound by general FOI law. (Showing both messages would be
confusing.)

Moves the EIR version of the message to make it equivalent to the
message about not being subject to legislation.