Fix TrackController#update with nil track_medium
Fail more gracefully when no track_medium param is supplied.

Raise an exception if no `track_medium` is supplied, instead of failing
with a 'no implicit conversion of nil into String' error when trying to
concatenate `nil` in to the error `String`.

Adds a spec for existing case of raising a `RuntimeError` when an
invalid `track_medium` param is supplied.

Fixes https://github.com/mysociety/alaveteli/issues/4871.