Prevent closed users signing in
We want to be able to retain a User record, but prevent them signing in.

Typically we ban users, but they can still sign in and perform a limited
set of actions. In some cases we want to completely prevent that. At the
moment, a developer will need to set `User#closed_at` to enact this, but
that's fine for our current use case of preventing access by a few
specific accounts.

Part of https://github.com/mysociety/alaveteli/issues/3148.