Show data from outgoing_message.raw_body on the admin edit page
This allows the admin to see and edit the unredacted copy of the message
from the raw_body attribute without indavertantly persisting what might
be a temporary redaction. Using a text_area_tag for this means that we
don't have to override how raw_body is persisted in the model.