Improve how accordion is drawn on admin_request#show page
Uses value from the current column - in line with the rest of the
display code - rather than calling outgoing_message.body directly. This
way if there are other columns we would like to display using the same
technique, we can more easily reuse it (or even extract it out to use
elsewhere).