Fix erroneously escaped HTML
At some point the omission links started to get escaped. I initially
thought it was introduced in e283db7, but it may have just been a rails
upgrade that change default escaping behaviour.

In any case, this fixes the problem by using a block to show extra
content when the text is truncated [1]. `omission` itself gets set to an
empty string so that we don't get a double `...`.

Fixes https://github.com/mysociety/alaveteli/issues/4905.

[1]
https://api.rubyonrails.org/v4.2.10/classes/ActionView/Helpers/TextHelper.html#method-i-truncate