Fix broken PublicBody factories
Using an `after(:create)` for the `tag_string` attribute prevents the
factories working as expected when using `FactoryBot.create`.

The `after(:create)` callback here doesn't persist the `tag_string`
modification, so if the record is reloaded during a spec then it is
effectively reset back to `nil`.

This commit adds traits for the particular cases, and then uses those
traits as attributes so that the named factories are still available.

The named factories have been deprecated in favour of traits.