Fix slow rendering of batch requests
The `embargo_form` partial was relying on `all_requests_created?` to
decide whether to disable the change embargo button. On large batches,
this meant iterating through every request in the batch.

Now that `sent_at` is sent after the batch sending is completed rather
than after the first request in the batch was sent (4af4ff4) we can use
`sent_at` to confirm that it is safe to change the embargo.

This adds `InfoRequestBatch#can_change_embargo?` as a nicer API method
to check whether its safe to perform the action.

Fixes https://github.com/mysociety/alaveteli/issues/4891.