Add IncomingMessage.unparsed
Adds a scope to find unparsed incoming messages. This will be used to
identify messages for background processing.

Adds a trait to build an unparsed `IncomingMessage`.

The spec implementation could be more concise…

    describe '.unparsed' do
      subject { described_class.unparsed }
      let!(:parsed) { FactoryBot.create(:incoming_message) }
      let!(:unparsed) { FactoryBot.create(:incoming_message, :unparsed) }
      it { is_expected.to match_array [unparsed] }
    end

…but this fails due to the existence of fixture data. Adding
`before { described_class.destroy_all }` doesn't solve the problem, as
that also destroys the sample data for the spec.