squash! Link to parent batch, from request header and sidebar
Also refactors the request subtitle in to separate partials for single
requests and batches.

The subtitle code was getting quite complex with the variations of
batch and whether to show admin links.

This fixes a translation using concatenation to build sentences [1].

Note that the partials now use a locals rather than relying on ivars
being set.

[1] https://mysociety.github.io/internationalization.html#dont-concatenate-strings-to-build-a-sentence