Handle new billing period webhook
Every month we receive a webhook letting us know that each customer has
moved to a new billing period. There's no action we want to take here,
so ignore it.

Explicitly tests cancellation as that happens under the same webhook
type (`customer.subscription.updated`). We _do_ want to get notified of
those, hence the test.

Also notifies us of trial end webhooks, as we don't really handle that
situation in the app very well, so we'd need to manually get in touch to
explain that the user needs to add a payment method etc.

Fixes https://github.com/mysociety/alaveteli-professional/issues/564.