Use helper for request hidden user explanation
This adds a helper for generating the hide request options, and adds the
distinction between the new state for the request and the “message” to
send to the user.

The reason for this is that we want to add message variations for common
offences, but still set the request to the same state after hiding it
and informing the user.

It would be more ideal to not have the duplication between state and
message when they are identical. I didn’t try to resolve this though, as
we intend to move these in to database-backed, admin-editable templates.