Fix bug in RawEmail#mail!
The implementation of `#mail!` in b479952129 is incorrect and can lead
to a strange result where `#mail` is returning `A`, even after calling
`#mail!` has returned `B`.

`#mail!` should return a fresh copy, but also _update the cache_ to this
latest copy, rather than returning the new copy and then letting the
cache return a stale version.