Rework Public Body tag scopes
This removes usages of the undocumented `AliasTracker` class as the
private interface has changed between Rails versions causing many spec
failures in our Rails 5 branches.

This has the added benefit of improving the performance of chained
`without_tag` calls, as used in the `PublicBody.is_requestable`. This
was causing a slowdown in the batch request sending.

Fixes #5044