Add PublicBodyChangeRequest#add_body_request?
A simpler way of telling updates from new authority requests than
directly checking for a public_body everywhere without creating
subclasses.

I've chosen to make a request to add a new body as the expectional case
as the class name sets up the expectation that we're going to receive
a request to change something that already exists.