Make the updated_before scope take outgoing_messages into account
The InfoRequest.updated_before scope is used solely by the 2
calculations within `stop_new_responses_on_old_requests` that determine
whether a request should be closed to new responses. Therefore rather
than "reopening" it when a followup is sent, this uses the followup
creation date when determining the request's age.