Be stricter about Integer conversion
When we attempt to guess an `InfoRequest` by a malformed email address,
we may pass a malformed ID to `_extract_id_hash_from_email`.

`to_i` will strip any letters from the end of a `String` that starts
with numerical values.

    '123ab'.to_i
    # => 123

This causes a transient spec failure when `FactoryBot` creates an
`InfoRequest` with the id `123` by chance, and this liberal conversion
results in us matching the request by `:id` instead of the expected
`:idhash`.

I've chosen to be strict here – rather than trying to parse out a
numeric value with a Regexp – because I think
`_extract_id_hash_from_email` should ether be able to extract the value
from the known format or not. We're introducing better guessing for
malformed email addresses in other commits, so a method like
`_guess_id_hash_from_email` is a more appropriate place to try to parse
out IDs from malformed addresses.

The two specs in this commit address two slightly different cases:

* malformed id: correct, but with character suffix
* broken id: completely unparsable ID

Fixes https://github.com/mysociety/alaveteli/issues/4996.