Refactor PublicBodyChangeRequestMailer
Extract method with flag parameter to separate methods [1].

The mailer shouldn't care about the state of a given
`PublicBodyChangeRequest`. This commit separates methods to comply with
SRP. Each method is simply responsible for rendering the given change
request in to a template.

Since `PublicBodyChangeRequest` holds the state of which type of change
request it is, _it_ should pick the correct mailer method to call.

Ideally we'd not even have this conditional [2], but its too much work
to extract out the separate classes of change request for now. A better
end result might look like:

    class AddPublicBodyRequest
      def send_message
        PublicBodyChangeRequestMailer.add_public_body(self)
      end
    end

    class UpdatePublicBodyRequest
      def send_message
        PublicBodyChangeRequestMailer.update_public_body(self)
      end
    end

[1] https://github.com/uohzxela/clean-code-ruby#dont-use-flags-as-function-parameters
[2] https://refactoring.guru/replace-conditional-with-polymorphism