Switch back to the geoip-database package
The geoip-database-contrib package attempts to download the legacy data
files which have now been removed from MaxMind's web servers.

The latest geoip-database package supplies the November 2018 version of
these files which should be safe to use until we can find a way to
ingest the v2 format files.

Connects to #5040