Add `current_path_with_locale` helper
This fixes an issue where the parameters weren't sanitized before being
reused in `url_for`. Now we strip the problematic parameters and ensure
we only ever return paths and not full URLs with protocol, host etc...

The full error message which was being outputted for many controller
and integration specs was:

`Attempting to generate a URL from non-sanitized request parameters! An
attacker can inject malicious data into the generated URL, such as
changing the host. Whitelist and sanitize passed parameters to be
secure.`