Redirect to the user admin page after updating comment visibility
It's the admin_user controller and that's what the spec expects to
happen so this seems like a better option adding a rails5? block to
try to call `redirect_back` with the user admin page as the
fallback_location param.

Fixes Rails 5 raising a deprecation warning for `redirect_to :back`