Use `delivery_status_will_change!` to tell Rails to update the attribute
Alas Rails 5.0 still tries to read the old value - and fails if it's
invalid - so we can't simplify our code.

But `save_changed_attribute` is no longer in the API so we need to use
the new `[name_of_attribute]_will_change!` method instead so that the
call to save retains the value we've set via `raw_write_attribute`.