Use `redirect_back` with a fallback location for Rails 5
Prevents Rails 5 from raising a deprecation warning for the use
of `redirect_to :back`.

As it seems reasonable to assume that the fallback location will only
be used rarely in practice, the redirect goes back to the admin user
index (list) page rather than attempting to calculate a user_id (as none
is supplied here) from an optional array of comment_ids.