Fix failing Users::SessionsController spec
The logger seems noisier under Rails 5 and additionally receives calls
to info with nil (multiple times) and a message about not rendering the
view - which breaks the original test.

Switching to using a spy instead to check that the expected message has
been received (works in Rails 4.2 so no special rails5? logic needed)