squash! Fix broken user profile updates spec
We've updated the capybara gem from 2.x to 3.x and the newer version no
longer normalises the whitespace when returning the text version of the
page source (so we end up with a newline char between HTML block
elements that was not there before). It is possible to restore the old
settings but as this ma be removed in future versions, the advice is to
fix the assertion instead.

See: https://stackoverflow.com/a/54672926