Pass JSON representation of payload when computing signatures in Rails 5
When Stripe computes the expected signature to make sure it matches (the
payload has not been altered in transit), it uses a JSON representation
of the Event data.

If we pass in the Hash (or a String representation of it), we'll end up
comparing a hash of e.g. '{"id" => "abc"}' with '{"id":"123"}' resulting
in the following error for all our signed webhook mocks:

Stripe::SignatureVerificationError:
  No signatures found matching the expected signature for payload