Update specs for changes in capybara 3.x
Capybara 3 no longer normalises whitespace so we need to update our
specs to compensate (only affects our Rails 5 branch as we're still
running v2.x against Rails 4).

Rather than force `normalize_ws: true` this fixes the expectations so
that we won't risk having to fix it again if this option is removed in
future verions. This means that the chnages for `have_link` are
consistent with our overall approach as it doesn't accept `normalize_ws'
as an option.

See: https://stackoverflow.com/a/54672926