Use allow rather than expect to stub call to #requestable_public_bodies
With Rails 5, this gets called multiple times (instead of the expected
single call) as the after_commit hook is being called by the spec and
I can't see a way to disabled it. As this is intended to be a test of
whether public_bodies which have already been sent the email are
excluded from the create_batch! email sending loop, it seems safe not
to verify that the requestable_public_bodies method is called.

Avoids raising error:

Failure/Error:
  expect(info_request_batch).to receive(:requestable_public_bodies).
    and_return([first_public_body])

  (#<InfoRequestBatch>).requestable_public_bodies(*(any args))
      expected: 1 time with any arguments
      received: 5 times with any arguments