Compare against bodies rather than draft.public_bodies
When we create the new batch, the controller deletes the draft.
Under Rails 5, this means we no longer have access to the draft's
public bodies as it tries to fetch a fresh copy using the relation
mapping and returns an empty array. (Previous versions of Rails appear
to have held the data in memory.)

As we already have a variable that holds the original array of the
public_bodies that was used to create the draft, we can just use this
for the comparison instead.

Fixes spec failure:

AlaveteliPro::InfoRequestBatchesController#create when everything is valid creates an info_request_batch
Failure/Error: expect(new_batch.embargo_duration).to eq draft.embargo_duration

  expected collection contained:  []
  actual collection contained:    [#<PublicBody id: 8, version: 1, last_edit_editor: "admin user", last_edit_comment: "Making an edit",...ts_overdue_count: nil, info_requests_visible_classified_count: nil, info_requests_visible_count: 0>]
  the extra elements were:        [#<PublicBody id: 8, version: 1, last_edit_editor: "admin user", last_edit_comment: "Making an edit",...ts_overdue_count: nil, info_requests_visible_classified_count: nil, info_requests_visible_count: 0>]