Update the check for a Stripe renewal to include latest_invoice
It looks as though Stripe are now including `latest_invoice` in the
`previous_attributes` data of renewal webhooks. Between 17:15:58 and
20:02:01 on 22/02/2019, the Event description changed from "[customer]
has started a new billing period" to "[customer]'s subscription has
changed" and includeing the new attribute. This behaviour appears to be
consistent across paid and discounted subscriptions.

This commit assumes that this is a permanent change and that we don't
have to look for renewals with and without `latest_invoice`.