Replace dynamic route with a workaround
Passing in a dynamic action is deprecated in Rails 5.1 (for security
reasons) and will be removed in 5.2.:

https://github.com/rails/rails/pull/23980

(We're seeing the deprecation notice in the latest build of 5.0.)

As we would like to keep this functionality, I've adopted a version of
the fix suggested - by the author of the code that introduced the
deprecation - in this issue thread in the Rails repo:

https://github.com/rails/rails/issues/27231#issuecomment-371066748

Essentially, this adds a filter to the allowable action name for safety
and uses :template as a proxy for :action. The main downside is that
themes that call `help_general_url` will need to pass in :template
instead of :action.

Also this caused (revealed?) a glitch which wasn't `/help/unhappy` route
- on its own without the optional `:url_title` part - wasn't being
recognised so that has also been updated to make :url_title optional