Make checking the file upload function an integration spec
As what we're trying to check here is that the uploaded file is accessed
and the contents used to create a preview of the information we're about
to load into the database when we press the "Dry run" button, this makes
more sense as an integration spec.

Also with the Rails 5 changes in handling parameters, the original spec
fails as it's harder to intercept the actual file object as it's no
longer a case of passing it in as a param. And our original stubbed the
file read in any case so was only testing that `read` was called, this
replacement checks that the content appears on screen as expected.