Allow send_error events to be treated as a special kind of send event
Needed for the change to InfoRequest#calculate_last_event_forming_initial_request
work as expected and to prevent introducing a bug in
InfoRequestEvent#is_request_sending?

The pre-existing event types related to message sending are divided into
'send' and 'resend' which relate to the original outgoing_message
(initial request) and 'send_followup' and 'resend_followup' which are
used for events relating to followup messages.

The new event type for 'send_error' is shared between the 2 types of
outgoing message. Rather than complicate the main 'send_error' code,
this commit adds some extra logic that checks message_type to
differentiate between events relating to the initial request and events
relating to followup messages. This prevents 'send_error' events for
followup messages unintentionally being treated as a change to the
request itself. (A send error for a followup message should not be
returned as the last event forming an initial request, but a send error
related to an initial request message should be treated as a request
sending event.)