Use theme_exists? to call theme method that loads the theme asset paths
Does this after `Rails.application.initialize!` as this is where the
initial asset paths are set, otherwise under Rails 5, our theme asset
paths end up as the last rather than the first search paths and fail to
override the defaults.

The theme method needs to look something like this:

```rb
def load_theme_assets
  # Prepend the asset directories in this theme to the asset path:
  ['stylesheets', 'images', 'javascripts'].each do |asset_type|
    theme_asset_path = File.join(Rails.root,
                                 'lib',
                                 'themes',
                                 THEME_NAME,
                                 'app',
                                 'assets',
                                 asset_type)

    Rails.application.assets.prepend_path theme_asset_path
  end
end
```

It cheats by calling the Sprockets method on `application.assets` (which
returns the Sprockets.environment`) rather than
`application.config.assets`, so we're bypassing SprocketsRails here. And
end up with conflicting information in the 2 asset arrays.

And have had to hack around with the theme code. So … probably not a
great solution but at least shows that 1) it is the asset_path order
and 2) it is fixable.