Prevent blank attributes being saved to translations
Mirror the `strip_attributes` rules for the parent class.

This is a particular problem for `PublicBody#name`, which when saved as
a blank String instead of `nil`, displays as the localised value, rather
than falling back to the default locale attribute. When used in the
admin interface as link text, it becomes unclickable.

Also fixes `Hash` style.

See https://github.com/mysociety/alaveteli/issues/5084.