Fix URL generation
An ArgumentError for non-sanitized parameters will be raised if
data is merged into request parameters and then used to generate
URLs with the `url_for` or `link_to` helpers.

We actually don't need to do this as these helpers already use the
current request parameters as a starting point for the URL generated.