Make sure cache files get the correct read permissions
In rails/rails@5d3b3cb Rails 5 updates `File.atomic_write` to only fall
back to using directory permissions for new files if the temporary file
has been created outside the current directory. Which for us has the
unintended effect of skipping the call to `chmod` which adds the
necessary read permissions for 'group' and 'others' that allow the web
server access to the cached files created by
`ApplicationController#foi_fragment_cache_write`.

In Rails 4.2 (with no pre-existing file) it works as follows:

A Tempfile is created with `Tempfile.new()`
After not finding a pre-existing file with a matching name to copy
permissions from, it looks at the default permissions of the containing
directory (which the owning process just created so this is exactly what
we want) and applies those to the new file.
https://apidock.com/rails/v4.2.9/File/atomic_write/class

In Rails 5.0 (again, no pre-existing file):

`File.atomic_write` now has a check that the destination directory is
not the same as the current directory before looking at directory
permissions. As in our case the temporary file is in the destination
directory, `old_stat` is nil and the Tempfile permissions (only readable
and writable by the owner) are retained.
https://apidock.com/rails/v5.0.0.1/File/atomic_write/class

This commit adds a call to `chmod` to add the missing permissions
(otherwise we end up with the `Tempfile.new` permissions which is
just read and write permissions for the owning user:
https://docs.ruby-lang.org/en/2.1.0/Tempfile.html#method-c-new).

Fixes #5241