Webhook state calculation
This has been heavily tested manually to workout the correct states for
the unknown webhooks which we've so far received.

I've attempted to write the conditionals in a way which wan't break but
can never be sure the webhook parameters will remain consistent and/or
other webhooks would be mis-categorised.