Prevent null bytes in attachment filename
While debugging a request I found:

    attrs[:filename]
    # => "Tender Loving Care Trust (Europe).pdf\u0000"

We've made a fix to the problem in the attachment body in f18c5b9.

This commit deletes null bytes from the filename attribute.

The deletion of the null bytes must happen before the extension
calculation, otherwise the spec fails with:

    expected: "Tender Loving Care Trust (Europe).pdf"
    got: "Tender Loving Care Trust (Europe).pdf.pdf"

We need a `try` as sometimes the filename being assigned is `nil` and
calculated by `FoiAttachment#ensure_filename!`. I didn't think a spec
was worth adding here, as the behaviour is covered by the
`ensure_filename!` spec

In 6c8490b we attributed a similar problem to Tnef. I'm not sure of the
cause of the error in this case.

Fixes https://github.com/mysociety/alaveteli/issues/5209.