Lock stripe-ruby-mock gem to '< 2.5.7'
Unless we upgrade from a 3.x to a 4.x version of the stripe gem
upgrading stripe-ruby-mock above 2.5.6 adds an undocumented dependency
on version 4.12.0+ of the stripe gem (where the Stripe::Discount object
was first added[1]). Untested assumption but it seems likely that we would
need to move to a newer Stripe API version to be able to upgrade the
stripe gem.

Avoids spec error:

NameError:
       uninitialized constant Stripe::Discount

[1] https://github.com/stripe/stripe-ruby/commit/160028ada110357c91740f26958dd7add399415f