Disable stale webhooks once we've created the new one
To avoid having multiple webhook endpoints, we should ideally delete
rather than just disable them (we should not really be expecting site
admins to delete things from their Stripe account on our behalf) but
to do that, we'd need to upgrade the stripe gem to at least v4.12.0
which is where the `delete` API call is introduced.

This disables the duplicate webhooks instead which at least prevents
them from continuing to listen for events. Having lots of stale webhooks
may cause a Stripe error when trying to create a new one:

  Stripe::InvalidRequestError
     You already reached the limit of 3 unique versions of test webhook
     endpoints. Please upgrade the versions of older test webhook
     endpoints to match the versions of newer test webhook endpoints.

(Not sure whether this is unique to the testing environment or a
reflection of the live behaviour.)