Don't error while sending not clarified reminders
The original error was raised (5dccf065) if the request doesn't have a
last response event. This makes some sense, as there wouldn't be
a request to clarify if the authority didn't respond.

bc743d9f adds a "public" qualifier to this test, as prominence was
introduced (4e2c4501) to `IncomingMessage`.

We now have a situation where this error can be raised when the request
is marked as `waiting_clarification`, but there's no last _public_
response due to the response prominence being set to hidden.

In the initial case, we should avoid allowing the user to classify as
`waiting_clarification`, but in the latter case a state of
`waiting_clarification` with a prominence of `hidden` is completely
valid.

For that reason, I think we should just skip that request when sending
reminders. As we won't log a `UserInfoRequestSentAlert` for the request,
it should get picked up later if the prominence is reset to normal.

Fixes https://github.com/mysociety/alaveteli/issues/3212.