Fix capistrano version at latest 2.x release
Alaveteli didn't upgrade to 3.x, which appears to have dropped the `-S`
flag we use in this guide. [1]

Personally I'd use `bundle exec cap…` to use the version bundled with
the rest of the gems Alaveteli uses, but this guide does say that you
may not actually have it fully installed on the deployment machine:

> checkout the Alaveteli repo (you need some of the files available
> locally even though you might not be running Alaveteli on this
> machine)

Fixes https://github.com/mysociety/alaveteli/issues/5386.

[1] https://groups.google.com/forum/?utm_medium=email&utm_source=footer#!msg/alaveteli-dev/ZFV2-1v2tok/yIX6tlpLBwAJ