Map TMail classes stored as YAML in the DB
TMail was dropped long ago as a result some requests are not viewable in
the admin because the InfoRequestEvent params YAML couldn't be parsed.

Now we're mapping these classes to an empty class. The attributes in the
YAML will be assign instance variables within the new object. If we need
access to the attributes in the future we can define a more complete
replacement class with accessors methods.

Fixes #5263