Replace TimeZone converter when deserialising YAML
Needed to fix issue in Rails 5.1 when attempting to deserialise YAML
generated with pre Rails 5. This might mean time zone information is
lost but only for objects stored in InfoRequestEvent#params_yaml

We don't appear to be using this at all.

Fixes #5463

Also include a previously identified mapping for a Postgres adapter
column. See #5264