Tweak change request edit form button
Reword button text to better describe the action performed. On the admin
summary page we use [Close | Close and respond | Add authority]. This
action also responds, so this commit uses that phrasing here.

It also aligns the button styles, by using `btn` instead of
`btn-primary`. This makes it easier to associate the action with the
button colour.