Add close without response to change request edit
The admin summary allows change requests to be closed without
responding to the user, or closed with a response to the user. The edit
form only allows the latter.

This is problematic because we send the edit form URL to the admin
contact address when we notify them of a new change request. If they
want to close without responding, they've got to go back to the admin
summary, find the change request, and close it that way.

This commit fixes https://github.com/mysociety/alaveteli/issues/5427 by
allowing change requests to be closed without responding through the
edit form.

I've used a link_to, because button_to submits the containing form, and
therefore submitting the response params.

This commit also cleans up some of the headings to reflect the new
behaviour.