Fix issue applying censor rule to binary data
The original implementation assumed the binary would be in ASCII-8BIT
encoding but this is not always the case. I'm unsure at this stage if
something has changed, such as how uncompressed data is extracted from
to the pdftk tool, or if this case has never been handled.

This change forces the data into ASCII-8BIT and back into the original
encoding regardless of what it initially was.