Limit number of Travis CI builds
We don't need to always test the next Ruby & Rails versions.

Ruby 2.3.3, which comes with Debian Stretch, is now only tested when
pushing to master and develop, not when a PR is opened - this version,
while supported, is old and Alaveteli installations should use a later
version of Ruby 2.3.x instead.

Ruby HEAD is now only tested on master and develop are pushed. This was
only ever meant to give an indication of future issues. We don't need
this to run on every PR.

The next version of Rails is now only tested for PR with a head branch
that begins with `rails`.