Always read the top-level description
Given a spec:

    describe 'foo/bar' do
      context 'a situation' do
        it 'renders as expected' do
          # …
          expect(rendered).to eq(read_described_template_fixture)
        end
      end
    end

Using `#description` would try to lookup a partial with the name 'a
situation'. We always describe the partial we're testing in the
top-level `describe` block (rather than a `context`) so use
`#top_level_description` [1] to ensure the correct template is read.

In some cases we may want separate fixture files, but at that point they
won't match the template file anyway so we may as well look them up
manually.

[1] https://stackoverflow.com/a/8745402/387558