Fix double-billing edge case
When a subscription goes `past_due` (most commonly after a failed
payment) we send an email suggesting the user update their card at
`/profile/subscriptions`. We only allow access to
`AlaveteliPro::SubscriptionsController` if the user has an _active_
subscription.

Prior to this commit, we defined "active" explicitly as the
`Stripe::Subscription#status` being `"active"`. This means that when a
sub moves to `past_due` the user can no longer access the controller to
add their new card.

We use the same check for `AlaveteliPro::PlansController`. So, when a
user goes `past_due` they can now re-subscribe to Pro and end up with
two subscriptions.

This commit:

* Adds the `AlaveteliPro::Subscription#past_due?` helper
* Adds the `AlaveteliPro::SubscriptionCollection#past_due` scope
* Allows `ProAccount#active?` to consider a `past_due` subscription as
  "active", in that we still consider a `past_due` sub as "OK" – we're
  assuming the payment problem is temporary, or rely on the
  auto-cancelling to clean up if the user becomes delinquent.