Consider all Stripe subscriptions "active"
Stripe::Customer#subscriptions [1] returns subscriptions that can have
one of many statuses [2].

However, it only includes the customer's _current_ subscriptions; it
does not include subscriptions that have been cancelled or that have
expired.

For the purposes of this access check, we only really care that the user
has a "current" subscription, even if its "past_due", "unpaid" or
whatever. We could include logic to better handle some of these cases,
but in the UI they would be handled _after_ the general access check.

Without this change, some users (e.g. those with a "trialing"
subscription) won't be able to get past some of the access checks, even
though they should be able to.

[1] https://stripe.com/docs/api/customers/object#customer_object-subscriptions
[2] https://stripe.com/docs/api/subscriptions/object#subscription_object-status