Rename ProAccount#active? to ProAccount#subscription?
Now that `#active?` is only checking whether a "current subscription"
[1] exists, "active" is quite misleading. An "active" subscription in
Stripe terms is one that is fully paid. However, there are other
statuses that are not "active" ("past_due", etc) that, in the context of
checking whether our customer can access certain areas of Pro, are
perfectly valid. We treat a "past_due" user the same as an "active" user
with respect to access – we just direct them to update their card
details to ensure payment can be taken.

Using `#subscription?` clarifies the intent here. "If the customer has a
current subscription, let them continue to whatever action we've
directed them to". We only disallow access to customers who once had a
subscription, but no longer do.

[1] https://stripe.com/docs/api/customers/object#customer_object-subscriptions