Highlight importance of saving OTP
We had an instance of a user needing their OTP reset and read out over a
video call. It doesn't happen frequently (this particular case is the
first that I know of), but if/when it does, it's extremely expensive (in
terms of time) to resolve. On reflection, we don't really highlight the
importance of saving the OTP, so this commit makes it much clearer.