Always db:seed on deploy
We don't add seeds regularly, but when we do they should get
automatically added as they're essential for the application.

Seeds should be idempotent, so this commit makes that explicit in the
comments for the file.

This also fixes https://github.com/mysociety/alaveteli/issues/4029, but
is more important for sites enabling Alaveteli Pro. If
`ENABLE_ALAVETELI_PRO` is set without the correct seeds in place, the
site will not work as expected.