Send InfoRequestBatch in the correct locale
Because `InfoRequestBatch.send_batches` is called in the background
(through `script/send-batch-requests`) the `AlaveteliLocalization` is
always set to the default.

This means that if the user creates an `InfoRequestBatch` in a
non-default locale, we try to match against an incorrect placeholder
salutation, resulting in all outgoing messages containing the default
placeholder ("Dear [Authority name]", if the default locale is EN).

Fortunately we store the last-used locale on `User`, so we can use that
to set the locale to use when sending batches. This commit makes the
assumption that a user doesn't change locale between creating the
`InfoRequestBatch` and it getting sent. This is pretty unlikely, so its
a reasonable fix for now.

A better fix would be to store `locale` on `InfoRequestBatch`, but that
would also require storing it on `DraftInfoRequestBatch` too (as if we
create from a draft, we just copy whatever text is in the draft). At
that point we may also want to think about storing locale per
`OutgoingMessage` – for now, too much to think about.

Fixes https://github.com/mysociety/asktheeu-theme/issues/103