Apply patch for CVE-2020-5267
> [CVE-2020-5267] Possible XSS vulnerability in ActionView
>
> There is a possible XSS vulnerability in ActionView's JavaScript
> literal escape helpers. Views that use the `j` or `escape_javascript`
> methods may be susceptible to XSS attacks.
>
> – https://groups.google.com/forum/#!topic/rubyonrails-security/55reWMM_Pg8

We use `j` in `app/views/general/_localised_datepicker.html.erb` for
displaying date pickers on the request search form.

Only Rails 5.2+ has received gem bumps.

The Rails 5.1 [1] and 5.2.4.1 [2] (the version before the patch)
implementations of `escape_javascript` are identical, so this patch
should be fine.

The commit message of the patch states:

> This commit escapes dollar signs and backticks to prevent JS XSS
> issues when using the `j` or `javascript_escape` helper
>
> – https://git.io/JvMcm

[1] https://git.io/JvMcI
[2] https://git.io/JvMcL