Don't show profile photos for banned users
Most users we ban are spam, so lets not show the profile photo. That way
we don't need need to do the extra step of removing it (time saver), and
it helps us build an archive of what a spam account looks like.

For banned users that aren't spam, they'll retain their profile photo
and it'll be displayed again when their ban is lifted.

Fixes https://github.com/mysociety/alaveteli/issues/5171.