Replace Ubuntu cosmic repo with focal repo
Cosmic has been decommissioned, so the site is not installable in
Ubuntu.

This commit replaces cosmic with focal. Focal is the next (20.04) LTS
release. It's not actually released at the time of committing, but it's
due out in 14 days from now so I think should be stable enough to rely
on now, rather than having to update it again when eoan disappears.

Also automatically removes old cosmic repo `.list` file. Users will need
to manually remove `cosmic` from `/etc/apt/preferences`, but as of this
commit, we use `preferences.d` for release-specific preferences so that
we can automate removal of old ones in future releases.